+++ src/Yaat.Client/Views/Radar/RadarView.ContextMenus.cs
@@ -45,7 +45,7 @@ public partial class RadarView
         var ac = mainVm.Aircraft.FirstOrDefault(a => a.Callsign == callsign);
         if (ac is not null)
         {
-            FlightPlanEditorManager.Open(ac, mainVm, TopLevel.GetTopLevel(this) as Window);
+            FlightPlanEditorManager.Open(ac, mainVm);
         }
     }